Separate databind test suite
